before building settings

+++ app/src/main/java/com/aritxonly/deadliner/MainActivity.kt
@@ -14,10 +14,15 @@ import android.view.WindowInsetsController
 import android.view.WindowManager
 import android.widget.Button
 import android.widget.ImageButton
+import android.widget.Toast
+import androidx.activity.result.ActivityResultLauncher
+import androidx.activity.result.contract.ActivityResultContracts
 import androidx.appcompat.app.AppCompatActivity
+import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.google.android.material.color.DynamicColors
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.google.android.material.floatingactionbutton.FloatingActionButton
 import java.time.LocalDateTime
 
@@ -29,6 +34,9 @@ class MainActivity : AppCompatActivity(), CustomAdapter.SwipeListener {
     private lateinit var settingsButton: ImageButton
     private val itemList = mutableListOf<DDLItem>()
     private lateinit var adapter: CustomAdapter
+    private lateinit var addDDLLauncher: ActivityResultLauncher<Intent>
+
+    private var pauseRefresh: Boolean = false
 
     override fun onCreate(savedInstanceState: Bundle?) {
         // 跟随主题色
@@ -57,13 +65,101 @@ class MainActivity : AppCompatActivity(), CustomAdapter.SwipeListener {
         // 设置 RecyclerView
         adapter = CustomAdapter(itemList, this)
         adapter.setSwipeListener(this)
+        // 设置单击监听器
+        adapter.setOnItemClickListener(object : CustomAdapter.OnItemClickListener {
+            override fun onItemClick(position: Int) {
+                val clickedItem = adapter.itemList[position]
+                pauseRefresh = true
+
+                val finishedString = if (clickedItem.isCompleted) {
+                    resources.getString(R.string.alert_edit_definish)
+                } else {
+                    resources.getString(R.string.alert_edit_finished)
+                }
+
+                // 选项数组
+                val options = arrayOf(
+                    resources.getString(R.string.alert_edit_modify),
+                    resources.getString(R.string.alert_edit_delete),
+                    finishedString
+                )
+
+                // 显示竖排按钮的 MaterialAlertDialog
+                MaterialAlertDialogBuilder(this@MainActivity)
+                    .setTitle(R.string.alert_edit_title)
+                    .setItems(options) { dialog, which ->
+                        when (which) {
+                            0 -> {
+                                // 修改操作
+                                val editDialog = EditDDLFragment(clickedItem) { updatedDDL ->
+                                    databaseHelper.updateDDL(updatedDDL)
+                                    adapter.updateData(databaseHelper.getAllDDLs())
+                                }
+                                editDialog.show(supportFragmentManager, "EditDDLFragment")
+                                pauseRefresh = false
+                            }
+                            1 -> {
+                                // 删除操作
+                                val item = adapter.itemList[position]
+                                databaseHelper.deleteDDL(item.id)
+                                adapter.updateData(databaseHelper.getAllDDLs())
+                                Toast.makeText(this@MainActivity, R.string.toast_deletion, Toast.LENGTH_SHORT).show()
+                                pauseRefresh = false
+                            }
+                            2 -> {
+                                // 标记为完成操作
+                                val item = adapter.itemList[position]
+                                item.isCompleted = !item.isCompleted
+                                databaseHelper.updateDDL(item)
+                                adapter.updateData(databaseHelper.getAllDDLs())
+                                if (item.isCompleted) {
+                                    Toast.makeText(
+                                        this@MainActivity,
+                                        R.string.toast_finished,
+                                        Toast.LENGTH_SHORT
+                                    ).show()
+                                } else {
+                                    Toast.makeText(
+                                        this@MainActivity,
+                                        R.string.toast_definished,
+                                        Toast.LENGTH_SHORT
+                                    ).show()
+                                }
+                            }
+                        }
+                    }
+                    .show()
+            }
+        })
+
         recyclerView.layoutManager = LinearLayoutManager(this)
         recyclerView.adapter = adapter
 
+        // 设置ItemTouchHelper
+        val itemTouchHelperCallback = object : ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.LEFT or ItemTouchHelper.RIGHT) {
+            override fun onMove(recyclerView: RecyclerView, viewHolder: RecyclerView.ViewHolder, target: RecyclerView.ViewHolder): Boolean {
+                return false
+            }
+
+            override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
+                when (direction) {
+                    ItemTouchHelper.LEFT -> adapter.onSwipeLeft(viewHolder.adapterPosition)
+                    ItemTouchHelper.RIGHT -> adapter.onSwipeRight(viewHolder.adapterPosition)
+                }
+            }
+        }
+        ItemTouchHelper(itemTouchHelperCallback).attachToRecyclerView(recyclerView)
+
+        addDDLLauncher = registerForActivityResult(ActivityResultContracts.StartActivityForResult()) { result ->
+            if (result.resultCode == RESULT_OK) {
+                // 更新数据
+                adapter.updateData(databaseHelper.getAllDDLs())
+            }
+        }
         // 添加新事件按钮
         addEventButton.setOnClickListener {
             val intent = Intent(this, AddDDLActivity::class.java)
-            startActivityForResult(intent, REQUEST_CODE_ADD_DDL)
+            addDDLLauncher.launch(intent)
         }
 
         settingsButton.setOnClickListener {
@@ -71,19 +167,45 @@ class MainActivity : AppCompatActivity(), CustomAdapter.SwipeListener {
         }
     }
 
+
+    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
+        super.onActivityResult(requestCode, resultCode, data)
+
+        if (requestCode == REQUEST_CODE_ADD_DDL && resultCode == RESULT_OK) {
+            // 刷新数据
+            adapter.updateData(databaseHelper.getAllDDLs())
+        }
+    }
+
     override fun onSwipeRight(position: Int) {
-        // 处理右滑删除事件
         val item = adapter.itemList[position]
-        databaseHelper.deleteDDL(item.id)
+        item.isCompleted = !item.isCompleted
+        databaseHelper.updateDDL(item)
         adapter.updateData(databaseHelper.getAllDDLs())
+        if (item.isCompleted) {
+            Toast.makeText(this, R.string.toast_finished, Toast.LENGTH_SHORT).show()
+        } else {
+            Toast.makeText(this, R.string.toast_definished, Toast.LENGTH_SHORT).show()
+        }
     }
 
     override fun onSwipeLeft(position: Int) {
-        // 处理左滑完成事件
-        val item = adapter.itemList[position]
-        item.isCompleted = true
-        databaseHelper.updateDDL(item)
-        adapter.updateData(databaseHelper.getAllDDLs())
+        pauseRefresh = true
+        MaterialAlertDialogBuilder(this)
+            .setTitle(R.string.alert_delete_title)
+            .setMessage(R.string.alert_delete_message)
+            .setNeutralButton(resources.getString(R.string.cancel)) { dialog, _ ->
+                adapter.notifyItemChanged(position) // 取消删除，刷新该项
+                pauseRefresh = false
+            }
+            .setPositiveButton(resources.getString(R.string.accept)) { dialog, _ ->
+                val item = adapter.itemList[position]
+                databaseHelper.deleteDDL(item.id)
+                adapter.updateData(databaseHelper.getAllDDLs())
+                Toast.makeText(this@MainActivity, R.string.toast_deletion, Toast.LENGTH_SHORT).show()
+                pauseRefresh = false
+            }
+            .show()
     }
 
     /**
@@ -137,14 +259,15 @@ class MainActivity : AppCompatActivity(), CustomAdapter.SwipeListener {
         return darkness < 0.5
     }
 
-    private val refreshInterval = 15000L
+    private val refreshInterval = 5000L
     private val handler = Handler(Looper.getMainLooper())
     private val refreshRunnable = object : Runnable {
         override fun run() {
             // 刷新数据
-            adapter.updateData(databaseHelper.getAllDDLs())
-            // 15秒后再次执行
-            handler.postDelayed(this, refreshInterval)
+            if (!pauseRefresh) {
+                adapter.updateData(databaseHelper.getAllDDLs())
+                handler.postDelayed(this, refreshInterval)
+            }
         }
     }
